cambios en las exportaciones y validaciones en el eliminar

+++ Controller/PeopleController.php
@@ -89,6 +89,9 @@ class PeopleController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
+		$persona_id = $this->Person->MeetingsPerson->find('first', array('conditions'=>array('MeetingsPerson.person_id' => $id)));
+		$this->set('persona_id', $persona_id);
+		debug($persona_id);
 		if (!$this->Person->exists($id)) {
 			throw new NotFoundException(__('Invalid person'));
 		}
@@ -122,10 +125,18 @@ class PeopleController extends AppController {
 			throw new NotFoundException(__('Invalid person'));
 		}
 		$this->request->onlyAllow('post', 'delete');
-		if ($this->Person->delete()) {
-			$this->Session->setFlash(__('The person has been deleted.'));
-		} else {
-			$this->Session->setFlash(__('The person could not be deleted. Please, try again.'));
+		
+		$persona_id = $this->Person->MeetingsPerson->find('first', array('conditions'=>array('MeetingsPerson.person_id' => $id)));
+		if($persona_id!=array()){
+			$this->Session->setFlash(__('La persona no se puede eliminar porque se encuentra asociada a una reunión.'));
+			return $this->redirect(array('action' => 'index'));
+			
 		}
-		return $this->redirect(array('action' => 'index'));
-	}}
+			if ($this->Person->delete()) {
+				$this->Session->setFlash(__('The person has been deleted.'));
+			} else {
+				$this->Session->setFlash(__('The person could not be deleted. Please, try again.'));
+			}
+			return $this->redirect(array('action' => 'index'));	
+	}
+}
